+++ changelog.md
@@ -5,6 +5,8 @@ ChangeLog
 * Rework editing of clients
 * Rework editing of notifications
 * Do not return current client
+* Fixed bug in client update
+* Optimized Styles
   
 1.2.6 (2020-08-16)
 -------------------------